Re-enabled the -render_indirect option to help with hardware headlight rendering on NVidia hardware. Real time shadowing will however be unavailable with indirect rendering. Significantly sped up high quality corona rendering. Decreased memory consumption during light data calculation for OpenGL style lighting.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@997 01821946-851b-0410-b7f4-f1477c70e57a

+++ arch/win32/ipx_udp.c
@@ -614,7 +614,7 @@ else {
 			break;
 		for (s2 = s; *s2 && *s2 != ','; s2++)
 			;
-		chk(ns=D2_ALLOC( s2 - s + 1));
+		chk (ns = D2_ALLOC ((unsigned int) (s2 - s + 1)));
 		memcpy(ns,s,s2-s);
 		ns[s2-s]='\0';
 		if (!queryhost(ns))
